requirements.txt
django
gunicorn
django-simple-captcha
python-decouple
django-admin-logs
mysqlclient
pytest
pytest-cov
junit-xml